feat: OCR implemented

+++ lib/app/models/ocr/ocr_model.dart
@@ -8,9 +8,9 @@ part 'ocr_model.g.dart';
 abstract class OcrModel with _$OcrModel {
   factory OcrModel({
     List<Items?>? items,
-    @JsonKey(name: 'tax_rate') int? taxRate,
-    @JsonKey(name: 'service_charge') int? serviceCharge,
-    @JsonKey(name: 'amount_paid') int? amountPaid,
+    @JsonKey(name: 'tax_rate') double? taxRate,
+    @JsonKey(name: 'service_charge') double? serviceCharge,
+    @JsonKey(name: 'amount_paid') double? amountPaid,
   }) = _OcrModel;
 
   factory OcrModel.fromJson(Map<String, dynamic> json) => _$OcrModelFromJson(json);
@@ -18,7 +18,7 @@ abstract class OcrModel with _$OcrModel {
 
 @unfreezed
 abstract class Items with _$Items {
-  factory Items({String? name, int? price, int? quantity}) = _Items;
+  factory Items({String? name, double? price, int? quantity}) = _Items;
 
   factory Items.fromJson(Map<String, dynamic> json) => _$ItemsFromJson(json);
 }
